Prevent crashing in supervisor_killable
`supervisor_killable` checks if a given Pid is part of the
given supersiors' tree via `supervisor:which_children`. It
may however happen that some of the given supervisors are
gone at the time of querying, either by themselves or a
sub-supervisor having been killed or crashed or exited by
natural means. In that case, `supervisor:which_children`
will raise an exit exception, and `havoc` will crash.